refactor(scripts): extract chain-text formatting helpers to lib/formatting.mjs
Moves nine pure formatting/sanitization functions from the 2 kLOC
scripts/lib.mjs monolith into scripts/lib/formatting.mjs as part of the

Extracted: slugify, formatLlmMarkdownText, classifyNativeName,
nativeNameQuality, nativeDisplayName, sanitizeChainText, stripUrls,
cleanDescription, deriveDescriptionFromNotes.

All nine are re-exported from scripts/lib.mjs so every existing importer
keeps its import path unchanged — pure code-motion with byte-identical
artifact output. The 98 %-coverage suite stays green (zero regression vs
the pre-change baseline).

Closes #510
